ihome_test/requirements.txt
alembic==0.9.4
amqp==2.2.2
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
pycryptodome==3.4.7
PyMySQL==0.9.3
python-alipay-sdk==1.4.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
qiniu==7.1.4
redis==2.10.5
requests==2.18.3
six==1.10.0
SQLAlchemy==1.1.12
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
WTForms==2.1
